+++ source/cfloat/matrices/sparse_row_compressed.c
@@ -94,7 +94,7 @@ jmtx_result jmtxc_matrix_crs_new(
     if (reserved_entries == 0)
     {
         reserved_entries = DEFAULT_RESERVED_ELEMENTS;
-        reserved_entries = reserved_entries < cols * rows ? reserved_entries : cols * rows;
+        reserved_entries = reserved_entries < ((uint64_t)cols * (uint64_t)rows) ? reserved_entries : ((uint64_t)cols * (uint64_t)rows);
     }
     if (allocator_callbacks == NULL)
     {
@@ -169,7 +169,7 @@ jmtx_result jmtxcs_matrix_crs_new(
     {
         return JMTX_RESULT_BAD_PARAM;
     }
-    if (reserved_entries > cols * rows)
+    if (reserved_entries > ((uint64_t)cols * (uint64_t)rows))
     {
         return JMTX_RESULT_BAD_PARAM;
     }
